+++ index.js
@@ -9,7 +9,6 @@ import messaging from '@react-native-firebase/messaging';
 
 
 messaging().setBackgroundMessageHandler(async remoteMessage => {
-    console.log('Message handled in the background!', remoteMessage);
   });
