+++ src/hooks/useAnalytics.ts
@@ -18,19 +18,9 @@ const useAnalytics = ({
   signUp,
   search,
 }: AnalyticsHandlerProps) => {
-  // useEffect(() => {
-  //   if (screenName) {
-  //     analytics().logScreenView({
-  //       screen_name: screenName,
-  //       screen_class: screenName,
-  //     });
-  //   }
-  // }, [screenName]);
-
   const track = {
     event: async (name: string, payload?: Record<string, any>) => {
       await analytics().logEvent(name, payload ?? {});
-      console.log("ho")
     },
     login: async (method: string) => {
       await analytics().logLogin({ method });
@@ -43,10 +33,10 @@ const useAnalytics = ({
     },
   };
 
-    if (functionName) track.event(functionName, data);
-    if (login) track.login(login);
-    if (signUp) track.signup(signUp);
-    if (search) track.search(search);
+  if (functionName) track.event(functionName, data);
+  if (login) track.login(login);
+  if (signUp) track.signup(signUp);
+  if (search) track.search(search);
 
   return { track };
 };
