+++ src/utils/helper.ts
@@ -69,33 +69,47 @@ const languages = [
   { label: 'हिंदी', value: 'hi' },
   { label: 'ગુજરાતી', value: 'gu' },
 ];
-
-const requestNotificationPermission = async (): Promise<boolean> => {
-  if (Platform.OS === 'android' && Platform.Version >= 33) {
+  const requestUserPermission = async () => {
     try {
-      const granted = await PermissionsAndroid.request(
-        PermissionsAndroid.PERMISSIONS.POST_NOTIFICATIONS,
-        {
-          title: 'Enable Notifications',
-          message:
-            'We’d like to show you notifications for alerts and updates.',
-          buttonNeutral: 'Later',
-          buttonNegative: 'Cancel',
-          buttonPositive: 'OK',
-        },
-      );
+      const granted: 'granted' | 'denied' | 'never_ask_again' =
+        await PermissionsAndroid.request(
+          PermissionsAndroid.PERMISSIONS.POST_NOTIFICATIONS,
+        );
 
-      return granted === PermissionsAndroid.RESULTS.GRANTED;
+      if (granted === PermissionsAndroid.RESULTS.GRANTED) {
+      } else {
+        console.log('Notification permission denied');
+      }
     } catch (error: any) {
-      console.error('Error requesting notification permission:', error);
-      return false;
+      console.error('Failed to request notification permission:', error);
     }
-  }
-};
+  };
+// const requestNotificationPermission = async (): Promise<boolean> => {
+//   if (Platform.OS === 'android' && Platform.Version >= 33) {
+//     try {
+//       const granted = await PermissionsAndroid.request(
+//         PermissionsAndroid.PERMISSIONS.POST_NOTIFICATIONS,
+//         {
+//           title: 'Enable Notifications',
+//           message:
+//             'We’d like to show you notifications for alerts and updates.',
+//           buttonNeutral: 'Later',
+//           buttonNegative: 'Cancel',
+//           buttonPositive: 'OK',
+//         },
+//       );
+
+//       return granted === PermissionsAndroid.RESULTS.GRANTED;
+//     } catch (error: any) {
+//       console.error('Error requesting notification permission:', error);
+//       return false;
+//     }
+//   }
+// };
 export {
   getAllUsers,
   checkUserExistsByEmail,
-  requestNotificationPermission,
+  requestUserPermission,
   updateUserInFirestore,
   languages,
 };
